moved search form styling to separate file since it will be reused
